+++ contrib/m_sipaf/backend/m_sipaf/migrations/versions/90f6e5531f7c_diagnostic_m_sipaf.py
@@ -29,7 +29,7 @@ def upgrade():
 
 def downgrade():
     if_exists = ""
-    # if_exists = "IF EXISTS"
+    if_exists = "IF EXISTS"
     op.execute(
         f"""
     DROP TABLE {if_exists} pr_sipaf.cor_diag_nomenclature_obstacle;